requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
typing-extensions>=4.8.0
